+++ src/archibald/models/__init__.py
@@ -1,10 +1,13 @@
 """Data models returned by operations."""
 
-from archibald.models.apply_edits_result import ApplyEditsResult, EditResultItem
+from archibald.models.apply_edits_result import ApplyEditsResult
+from archibald.models.attachments_result import AddAttachmentsResult
+from archibald.models.edit_result_item import EditResultItem
 from archibald.models.fields_result import FieldsResult
 from archibald.models.query_result import QueryResult
 
 __all__ = [
+    "AddAttachmentsResult",
     "ApplyEditsResult",
     "EditResultItem",
     "FieldsResult",